Allow very long {items} lists in fzy#start()
When a list of strings is passed to fzy#start(), internally the strings
are concatenated with '\n', passed to printf, and then executed in the
shell using the -c option.

When long arguments are passed to the shell like:

  /bin/bash -c 'printf "a very long argument..." | fzy'

the shell will print an error:

  bash: /bin/bash: Argument list too long

Further reading:
  https://stackoverflow.com/a/11475732
  https://www.in-ulm.de/~mascheck/various/argmax
  https://unix.stackexchange.com/a/120842
  https://stackoverflow.com/a/18647755

The maximum size seems to be 128KiB (131072 bytes) on Linux.

To avoid this, check the length of the printf-argument and if the string
is too long, write the items to a temporary file and pass this file to
fzy.

TODO: use FIFOs on systems where 'mkfifo' is available to avoid writing
to disk.