Pass list through FIFO to fzy on Linux/Unix
When "mkfifo" is provided by the OS, the list is written to a named pipe
and passed as stdin to fzy. This avoids unnecessary writes to disk. On
systems that do not provide "mkfifo", we write the items to a temporary
file on disk and pass the file as stdin to fzy. In either case, there's
no limit on the length of the list.

Currently a FIFO is used only for fzy input. The output of fzy (the
selected item) is always written to a file on disk. Using two FIFOs for
both input and output isn't trivial (TODO).