Use values defined in g:fzy before default values
All entries in the {options} argument of fzy#start() are optional.

vim-fzy checks the dictionary g:fzy for the keys that haven't been
passed to fzy#start(). If a key is set in g:fzy, it will be used. The
default values defined by the plugin are used as a fall back when
neither passed to fzy#start() nor defined in g:fzy.

Plugins using vim-fzy can now omit, for example, the "height" entry and
let users set it in their vimrc.